## Summary
The org dashboard view can be really slow for some orgs with lots of services and notifications. The queries that power this view haven't been optimised much, and so we can make 100s or 1000s of round trips to the DB. Let's refactor some of the supporting logic to make queries in bulk so that we have fewer round trips, and hopefully (much) faster page loads.

With the test case provided below, on my machine we go from ~1250 DB queries to 11, and from 8.6s to 0.3s

This will also speed up the nightly billing fact generation task (create-nightly-billing-for-day). Over the last 7 days in prod it's taken approx 300s each time.

On my machine, with the local test data from below, on `main` it takes 18 seconds and on this branch it takes <1 second. (the log messages don't have any decimal precision, so it comes back as 0 seconds).

## Testing locally
I've added a script to generate an organisation with 100 services and 5,000,000 notifications across them, so that the org dashboard view loads slowly locally.

1) Run: `flask command generate-bulktest-data -u <YOUR_LOCAL_USER_ID>`. This runs pretty slowly - expect to kick it off and walk away for 30 minutes.

    <details><summary>It takes a while...</summary>

    ```
    ~/work/gds/notifications-api [SW-reduce-org-dashboard-queries]$ flask command generate-bulktest-data -u 0c38e85c-b037-4dd8-ae12-09ab2b84c373
    [0.00]: Sending org to DB...
    [0.05]: Done.
    [0.05]: Building services...
    [0.05]: Sending services to DB...
    [0.09]: Done.
    [0.09]: Building templates...
    [0.11]: Sending templates to DB...
    [0.11]: Done.
    [0.11]: Building 5,000,000 notifications...
    [0.18]: Building batch 1...
    [28.20]: Sending batch 1 to DB...
    [370.70]: Done.
    [370.70]: Building batch 2...
    [399.35]: Sending batch 2 to DB...
    
    [920.54]: Done.
    [920.54]: Building batch 3...
    [948.87]: Sending batch 3 to DB...
    [1578.88]: Done.
    [1578.88]: Building batch 4...
    [1607.78]: Sending batch 4 to DB...
    [2292.68]: Done.
    [2292.68]: Building batch 5...
    [2320.96]: Sending batch 5 to DB...
    [3081.49]: Done.
    [3081.49]: Committing...
    [3081.53]: Finished.
    ```
    </details>

    <details><summary>To delete all(?) the data afterwards</summary>

    ```
    delete from notifications where client_reference LIKE 'BULKTEST%';
    delete from notification_history where client_reference LIKE 'BULKTEST%';
    delete from template_redacted where "template_id" IN (select "id" from templates where "name" like 'BULKTEST%');
    delete from templates where "name" like 'BULKTEST%';
    delete from templates_history where "name" like 'BULKTEST%';
    delete from service_permissions where service_id IN (select "id" from services where "name" like 'BULKTEST%');
    delete from permissions where service_id IN (select "id" from services where "name" like 'BULKTEST%');
    delete from user_to_service where service_id in  (select "id" from services where "name" like 'BULKTEST%');
    delete from service_sms_senders where service_id in  (select "id" from services where "name" like 'BULKTEST%');
    delete from annual_billing where service_id in  (select "id" from services where "name" like 'BULKTEST%');
    delete from services where "name" like 'BULKTEST%';
    delete from organisation where "name" like 'BULKTEST%';
    ```
    </details>
2) I've pushed up a commit 7c9a6aeb17feecef9fced26a10a7b7d5603a7f0b on a separate branch that adds a profiler and flask-debugtoolbar, as well as a fake view that replicates the org services-with-usage endpoint used by the org view dashboard.
    `git checkout main`
    `git cherry-pick 7c9a6aeb17feecef9fced26a10a7b7d5603a7f0b`
    `make bootstrap`
3) Open your browser and navigate to http://localhost:6011/organisations/BULKTEST_ORG_ID_HERE/services-with-usage-html?year=2022. Check the toolbar on the side and record the time taken and the number of queries executed.
4) Put your `main` branch back to normal: `git reset --hard main` / `git reset --hard head^` to remove the cherry-picked commit
5) Checkout this branch, `git checkout SW-reduce-org-dashboard-queries` and cherry-pick the profiling commit again:  `git cherry-pick 7c9a6aeb17feecef9fced26a10a7b7d5603a7f0b`
6) Run step 3 again. Compare the time taken and queries executed.

Note: the numbers on the dashboard will feel 'wrong', because they are - we haven't built historical billing data. This could be done by running `create_nightly_billing_for_day(date)` for all dates in the range that the command generates notifications within. However this doesn't, as far as I can see, affect the performance of the dashboard - just the accuracy of the stats.

## Comparison

| Before | After |
|---|---|
| <img width="215" alt="image" src="https://user-images.githubusercontent.com/2920760/206703262-9f5e36c3-cc8c-4b03-99fe-a5fc16201dea.png"> | <img width="209" alt="image" src="https://user-images.githubusercontent.com/2920760/206703323-2881a589-e2f7-42d4-9c01-44ed568ab42a.png"> |