@samuelhwilliams 2022-12-13
> I couldn't easily get my head around the second commit and it got too late for me to start doing it. It probably isn't that crazy but definitely touches some of our more complex code and I would need to do a fair bit of sqlalchemy docs reading. If you want to move this pull request through quicker, I'd be open to moving the second commit to a separate PR or dropping it (assuming that it is not the deal breaker commit that really reduces the bulk of the time). Such complex code of ours is hard to review and I'm not sure that your changes in that commit make it any easier (not that I would necessarily expect them to but if the performance benefits aren't drastic for the commit, I'm OK to not include it for that reason).

It should be possible to drop out the 2nd commit and do it as a separate PR, though we'd be "putting back" a fair number of queries (we'd continue to make 3n rather than 3 queries, where n is the number of services). I think it would be nice to keep in. I've given it another pass and I had some unnecessary separation that made it look like a bigger change than it is. If you look again, hopefully it appears less troublesome now 🤞 

> I also would like to run your new test test_get_organisation_services_usage_limit_queries_executed by itself against the main branch to see what the value was before as well. I might try and find 2 mins to do that tomorrow.

By all means. Hopefully this should be as simple as cherry-picking the commit onto `main` and running/investigating it with `pytest -vvvs -k test_get_organisation_services_usage_limit_queries_executed` or similar. For the 5 services case, I get 26 queries, and for the 10 services case, I get 46 queries.