@samuelhwilliams 2022-12-14
The failing test is unrelated to this PR (caused by enabling sqlalchemy query recording during tests, and this one single test is triggering some edge case inside sqlalchemy/psycopg2). Looking into ways around it, but shouldn't stop any further reviewing of this PR... 🤦